retail-data-warehouse-main/requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
python-dotenv==1.0.0